+++ mantidimaging/gui/ui/roi_form_widget.ui
@@ -219,7 +219,7 @@
   <customwidget>
    <class>ROIPropertiesTableWidget</class>
    <extends>QWidget</extends>
-   <header>mantidimaging.gui.windows.spectrum_viewer.view</header>
+   <header>mantidimaging.gui.widgets.spectrum_widgets.roi_form_widget</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
